+++ src/main/kotlin/app/revanced/patches/backdrops/misc/pro/annotations/ProUnlockCompatibility.kt
@@ -3,5 +3,5 @@ package app.revanced.patches.backdrops.misc.pro.annotations
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
 
-@Compatibility([Package("com.backdrops.wallpapers")])
+@Compatibility([Package("com.backdrops.wallpapers", arrayOf("4.52"))])
 internal annotation class ProUnlockCompatibility