+++ src/main/kotlin/app/revanced/patches/iconpackstudio/misc/pro/annotations/UnlockProCompatibility.kt
@@ -5,5 +5,4 @@ import app.revanced.patcher.annotation.Package
 
 @Compatibility([Package("ginlemon.iconpackstudio")])
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
 internal annotation class UnlockProCompatibility