+++ src/main/kotlin/app/revanced/patches/reddit/ad/general/annotations/GeneralAdsCompatibility.kt
@@ -5,9 +5,18 @@ import app.revanced.patcher.annotation.Package
 
 @Compatibility(
     [Package(
-        "com.reddit.frontpage", arrayOf("2022.43.0", "2023.05.0")
+        "com.reddit.frontpage", arrayOf(
+            "2021.45.0", 
+            "2022.43.0", 
+            "2023.05.0",
+            "2023.06.0",
+            "2023.07.0", 
+            "2023.07.1",
+            "2023.08.0",
+            "2023.09.0",
+            "2023.09.1"
+        )
     )]
 )
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
 internal annotation class GeneralAdsCompatibility