+++ src/main/kotlin/app/revanced/patches/twitch/ad/video/annotations/VideoAdsCompatibility.kt
@@ -3,8 +3,20 @@ package app.revanced.patches.twitch.ad.video.annotations
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
 
-@Compatibility([Package("tv.twitch.android.app")])
+@Compatibility(
+    [
+        Package(
+            "tv.twitch.android.app", arrayOf(
+                "14.3.3",
+                "14.4.0",
+                "14.5.0", 
+                "14.5.2",
+                "14.6.0",
+                "14.6.1"
+            )
+        )
+    ]
+)
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
 internal annotation class VideoAdsCompatibility