+++ src/main/kotlin/app/revanced/patches/youtube/misc/debugging/annotations/DebuggingCompatibility.kt
@@ -5,5 +5,4 @@ import app.revanced.patcher.annotation.Package
 
 @Compatibility([Package("com.google.android.youtube")])
 @Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
 internal annotation class DebuggingCompatibility